Refactor: input 노드를 추가하고, 삭제하는 모듈 생성
